Разбивает LoginScreen на дополнительные компоненты
