+++ src/components/login-screen/login-form/login-form.tsx
@@ -1,0 +1,73 @@
+import { FormEvent, useEffect, useRef } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { AppRoute, AuthorizationStatus } from '../../../const';
+import { loginAction } from '../../../store/api-actions';
+import { getAuthorizationStatus } from '../../../store/user-process/user-process.selectors';
+import { useAppDispatch, useAppSelector } from '../../hooks';
+
+function LoginForm(): JSX.Element {
+  const dispatch = useAppDispatch();
+  const loginRef = useRef<HTMLInputElement | null>(null);
+  const passwordRef = useRef<HTMLInputElement | null>(null);
+
+  const authorizationStatus = useAppSelector(getAuthorizationStatus);
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    if (authorizationStatus === AuthorizationStatus.Auth) {
+      navigate(AppRoute.Main, { replace: true });
+    }
+  }, [authorizationStatus, navigate]);
+
+  const handleSubmit = (evt: FormEvent<HTMLFormElement>) => {
+    evt.preventDefault();
+
+    if (loginRef.current !== null && passwordRef.current !== null) {
+      dispatch(
+        loginAction({
+          login: loginRef.current.value,
+          password: passwordRef.current.value,
+        })
+      );
+    }
+  };
+
+  return (
+    <form
+      className="login__form form"
+      onSubmit={handleSubmit}
+      action="#"
+      method="post"
+    >
+      <div className="login__input-wrapper form__input-wrapper">
+        <label className="visually-hidden">E-mail</label>
+        <input
+          className="login__input form__input"
+          type="email"
+          name="email"
+          placeholder="Email"
+          ref={loginRef}
+          required
+        />
+      </div>
+      <div className="login__input-wrapper form__input-wrapper">
+        <label className="visually-hidden">Password</label>
+        <input
+          className="login__input form__input"
+          type="password"
+          pattern="(?=.*\d)(?=.*[a-zA-Z]).{2,}"
+          title="Contains one letter and one digit"
+          name="password"
+          placeholder="Password"
+          ref={passwordRef}
+          required
+        />
+      </div>
+      <button className="login__submit form__submit button" type="submit">
+        Sign in
+      </button>
+    </form>
+  );
+}
+
+export default LoginForm;
