+++ src/pages/login-screen/login-screen.tsx
@@ -1,121 +1,33 @@
-import { FormEvent, useEffect, useRef, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { Helmet } from 'react-helmet-async';
-import { Link, useNavigate } from 'react-router-dom';
-import Logo from '../../components/ui/logo/logo';
-import { AppRoute, AuthorizationStatus, Cities } from '../../const';
-import { useAppDispatch, useAppSelector } from '../../hooks';
-import { loginAction } from '../../store/api-actions';
-import { changeCity } from '../../store/app-process/app-process.slice';
-import { getAuthorizationStatus } from '../../store/user-process/user-process.selectors';
+import Locations from '../../components/login-screen/locations/locations';
+import LoginForm from '../../components/login-screen/login-form/login-form';
+import Header from '../../components/ui/header/header';
+import { Cities } from '../../const';
 
 function LoginScreen(): JSX.Element {
-  const dispatch = useAppDispatch();
-
   const randomCityKey =
     Object.keys(Cities)[Math.floor(Math.random() * Object.keys(Cities).length)];
 
   const [randomCity, setRandomCity] = useState(randomCityKey);
-  const authorizationStatus = useAppSelector(getAuthorizationStatus);
-  const navigate = useNavigate();
 
   useEffect(() => {
     setRandomCity(randomCityKey);
   }, [randomCityKey]);
-
-  useEffect(() => {
-    if (authorizationStatus === AuthorizationStatus.Auth) {
-      navigate(AppRoute.Main, { replace: true });
-    }
-  }, [authorizationStatus, navigate, dispatch]);
-
-  function handleCityClick(city: string) {
-    dispatch(changeCity({ city }));
-  }
-
-  const loginRef = useRef<HTMLInputElement | null>(null);
-  const passwordRef = useRef<HTMLInputElement | null>(null);
-
-  const handleSubmit = (evt: FormEvent<HTMLFormElement>) => {
-    evt.preventDefault();
-
-    if (loginRef.current !== null && passwordRef.current !== null) {
-      dispatch(
-        loginAction({
-          login: loginRef.current.value,
-          password: passwordRef.current.value,
-        })
-      );
-    }
-  };
 
   return (
     <div className="page page--gray page--login">
       <Helmet>
         <title>6 cities :: Login</title>
       </Helmet>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Logo />
-            </div>
-          </div>
-        </div>
-      </header>
-
+      <Header />
       <main className="page__main page__main--login">
         <div className="page__login-container container">
           <section className="login">
             <h1 className="login__title">Sign in</h1>
-            <form
-              className="login__form form"
-              onSubmit={handleSubmit}
-              action="#"
-              method="post"
-            >
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">E-mail</label>
-                <input
-                  className="login__input form__input"
-                  type="email"
-                  name="email"
-                  placeholder="Email"
-                  ref={loginRef}
-                  required
-                />
-              </div>
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">Password</label>
-                <input
-                  className="login__input form__input"
-                  type="password"
-                  pattern="(?=.*\d)(?=.*[a-zA-Z]).{2,}"
-                  title="Contains one letter and one digit"
-                  name="password"
-                  placeholder="Password"
-                  ref={passwordRef}
-                  required
-                />
-              </div>
-              <button
-                className="login__submit form__submit button"
-                type="submit"
-              >
-                Sign in
-              </button>
-            </form>
+            <LoginForm />
           </section>
-          <section className="locations locations--login locations--current">
-            <div className="locations__item">
-              <Link
-                onClick={() => handleCityClick(randomCity)}
-                className="locations__item-link"
-                to={AppRoute.Main}
-              >
-                <span>{randomCity}</span>
-              </Link>
-            </div>
-          </section>
+          <Locations randomCity={randomCity} />
         </div>
       </main>
     </div>
